use new temp cluster for mija and aqara
